feat: board.js 컴포넌트 추가&적용
